Image popup js starting point
This exercise is being done to have on click event listeners attached
to images in which the image click will then show the full size render
of the thumbnail beneath it.
The code is trying to be optimized with functions to limit code
rewriting. Just passing a parameter to the same function to get from an
array the image and its comment at a certain index.